// src/shared/lib/generated/prisma/enums.ts
export const PaymentStatus = {
  PENDING: 'PENDING',
  SUCCEEDED: 'SUCCEEDED',
  FAILED: 'FAILED',
  REFUNDED: 'REFUNDED'
} as const

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus]
// ...
export const PaymentProvider = {
  STRIPE: 'STRIPE',
  JAZZCASH: 'JAZZCASH',
  COD: 'COD'
} as const

export type PaymentProvider = (typeof PaymentProvider)[keyof typeof PaymentProvider]